added unrolling for gemv function

+++ src/cuda_standard_kernel/cuda_device_helpers.cuh
@@ -21,8 +21,8 @@ template<typename T, size_t THREADS_PER_BLOCK, size_t WARP_SIZE = 32>
 __device__ __forceinline__ void block_warp_reduction(T value, const uint tidx, T *smmem) {
     T local_sum{warp_reduction<T, WARP_SIZE>(value)};
     if (THREADS_PER_BLOCK > WARP_SIZE) {
-        const uint c_warp{tidx / WARP_SIZE};
-        const uint c_t_pos{tidx % WARP_SIZE};
+        const size_t c_warp{tidx / WARP_SIZE};
+        const size_t c_t_pos{tidx % WARP_SIZE};
 
         if (c_t_pos == 0) smmem[c_warp] = local_sum;
